+++ src/ResourceManager/Profile/Commands.Profile/StartupScripts/AzureRmError.ps1
@@ -3,13 +3,17 @@ function Write-InstallationCheckToFile
     Param($installationchecks)
     if (Get-Module Az.Profile -ListAvailable)
     {
-        Write-Warning "Both Az and AzureRM modules were detected on your machine. Az and AzureRM module cannot be run side-by-side, please run 'Uninstall-AzureRm' to remove all AzureRm modules from your machine. More information can be found here: aka.ms/azps-migration-guide"
+        Write-Warning "Both Az and AzureRM modules were detected on your machine. Az and AzureRM module cannot be run side-by-side, please run 'Uninstall-AzureRm' to remove all AzureRm modules from your machine. More information can be found here: https://aka.ms/azps-migration-guide"
     }
 
     $installationchecks.Add("AzureRmSideBySideCheck","true")
     try
     {
-        Remove-Item -Path $pathToInstallationChecks
+        if (Test-Path $pathToInstallationChecks)
+        {
+            Remove-Item -Path $pathToInstallationChecks
+        }
+        
         New-Item -Path $pathToInstallationChecks -ItemType File -Value ($installationchecks | ConvertTo-Json)
     }
     catch
@@ -21,27 +25,13 @@ function Write-InstallationCheckToFile
 if (!($env:SkipAzInstallationChecks -eq "true"))
 {
     $pathToInstallationChecks = Join-Path (Join-Path $HOME ".Azure") "AzInstallationChecks.json"
+    $installationchecks = @{}
     if (!(Test-Path $pathToInstallationChecks))
     {
-        if (Get-Module Az.Profile -ListAvailable)
-        {
-            Write-Warning "Both Az and AzureRM modules were detected on your machine. Az and AzureRM module cannot be run side-by-side, please run 'Uninstall-AzureRm' to remove all AzureRm modules from your machine. More information can be found here: aka.ms/azps-migration-guide"
-        }
-
-        $hashtable = @{"AzureRmSideBySideCheck"="true"}
-        try 
-        {
-            New-Item -Path $pathToInstallationChecks -ItemType File -Value ($hashtable | ConvertTo-Json)
-        }
-        catch 
-        { 
-            Write-Verbose "Installation checks failed to write to file." 
-        }
+        Write-InstallationCheckToFile $installationchecks
     }
-
     else
     {
-        $installationchecks = @{}
         try
         {
             ((Get-Content $pathToInstallationChecks) | ConvertFrom-Json).PSObject.Properties | Foreach { $installationchecks[$_.Name] = $_.Value }
@@ -60,6 +50,6 @@ if (!($env:SkipAzInstallationChecks -eq "true"))
 
 if (Get-Module Az.profile)
 {
-    Write-Warning "Az.Profile already loaded. Az and AzureRM module cannot be run side-by-side, please run 'Uninstall-AzureRm' to remove all AzureRm modules from your machine. More information can be found here: aka.ms/azps-migration-guide"
-    throw "Az.Profile already loaded. Az and AzureRM module cannot be run side-by-side, please run 'Uninstall-AzureRm' to remove all AzureRm modules from your machine. More information can be found here: aka.ms/azps-migration-guide"
+    Write-Warning "Az.Profile already loaded. Az and AzureRM module cannot be run side-by-side, please run 'Uninstall-AzureRm' to remove all AzureRm modules from your machine. More information can be found here: https://aka.ms/azps-migration-guide"
+    throw "Az.Profile already loaded. Az and AzureRM module cannot be run side-by-side, please run 'Uninstall-AzureRm' to remove all AzureRm modules from your machine. More information can be found here: https://aka.ms/azps-migration-guide"
 }
